pyproject.toml
[tool.poetry]
name = "FastApi BackEND"
version = "0.1.0"
description = ""
authors = ["PauloJunior65 <paulojunior051996@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.31"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
sqlmodel = "^0.0.19"
alembic = "^1.13.1"
pymysql = "^1.1.1"
aioredis = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.4.10"
httpx = "^0.27.0"

[tool.ruff]
line-length = 320
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=./ -vv'
post_test = 'coverage html'
migrations = 'alembic revision --autogenerate -m'
migrate = 'alembic upgrade head'

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
